Modularized the project
