+++ main.py
@@ -1,0 +1,6 @@
+from game import Game
+
+# Main entry point to run the game
+if __name__ == "__main__":
+    game = Game()
+    game.play()
